ATRIL BACKPORTS: tiff: fix compile warning
commit 85ee1df4499969dd3c6dee3bd032a18d8bc34999
Author: raveit65 <chat-to-me@raveit.de>
Date:   Sun Jul 3 13:30:09 2016 +0200

    tiff: fix compile warning

    taken from:
    https://git.gnome.org/browse/evince/commit/?id=2059456
    https://git.gnome.org/browse/evince/commit/?id=00b5e55

+++ backend/tiff/tiff2ps.c
@@ -1131,7 +1131,7 @@ PS_Lvl2page(TIFF2PSContext* ctx, TIFF* tif, uint32 w, uint32 h)
 	buf_data = (unsigned char *)_TIFFmalloc(chunk_size);
 	if (!buf_data) {
 		TIFFError(ctx->filename, "Can't alloc %u bytes for %s.",
-			chunk_size, tiled_image ? "tiles" : "strips");
+			(uint) chunk_size, tiled_image ? "tiles" : "strips");
 		return(FALSE);
 	}
